Expand >=media-sound/easyeffects-6.0.0 to specific version numbers
It seems ranges do not work on p.g.o. yet. The curent list includes both
versions already released upstream (up to and including 6.0.2) as well
as some not released yet, so that we do not have to update this list
every time a patch release comes out - but even so, let's hope we will
not have to wait for gtk4 for too long!

Signed-off-by: Marek Szuba <marecki@gentoo.org>