Ignore app-crypt/libu2f-host::slackbuilds
This one is a weird one - Repology thinks SlackBuild contains 1.1.11 yet
following their own link to
https://slackbuilds.org/repository/14.2/libraries/libu2f-host/ shows
version 1.1.10 like everyone else. Hopefully it will go away at some point.

For the record, the latest version published upstream is 1.1.10 and with
libu2f-host having been deprecated in favour of libfido2, it is
extremely unlikely for 1.1.11 to materialise.

Signed-off-by: Marek Szuba <marecki@gentoo.org>